+++ index.html
@@ -6,7 +6,10 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>tic-tac-toe</title>
 		<link rel="stylesheet" href="style.css" />
-		<script defer src="app.js"></script>
+		<script defer type="module" src="displayController.js"></script>
+		<script defer type="module" src="gameBoard.js"></script>
+		<script defer type="module" src="gameController.js"></script>
+		<script defer type="module" src="game.js"></script>
 	</head>
 	<body>
 		<header class="primary-header">Tic Tac Toe</header>
@@ -36,7 +39,7 @@
 					<button class="field-btn"><p></p></button>
 					<button class="field-btn"><p></p></button>
 				</section>
-				<button type = "button" id ="reset">Start Over</button>
+				<button type="button" id="reset">Start Over</button>
 			</div>
 		</main>
 	</body>
